+++ src/Barbies/Internal/Wrappers.hs
@@ -6,12 +6,11 @@ module Barbies.Internal.Wrappers
   ( Barbie(..)
   ) where
 
+import Barbies.Internal.Applicative
 import Barbies.Internal.Constraints
 import Barbies.Internal.Dicts
 import Barbies.Internal.Functor
 import Barbies.Internal.Traversable
-import Barbies.Internal.Product
-import Barbies.Internal.ProductC
 
 import Data.Kind (Type)
 import Data.Semigroup (Semigroup, (<>))
@@ -21,7 +20,7 @@ import Prelude hiding (Semigroup, (<>))  -- ghc < 8.2
 -- | A wrapper for Barbie-types, providing useful instances.
 newtype Barbie (b :: (k -> Type) -> Type) f
   = Barbie { getBarbie :: b f }
-  deriving (FunctorB, ProductB, ProductBC)
+  deriving (FunctorB, ApplicativeB)
 
 -- Need to derive it manually to make GHC 8.0.2 happy
 instance ConstraintsB b => ConstraintsB (Barbie b) where
@@ -32,12 +31,12 @@ instance TraversableB b => TraversableB (Barbie b) where
   btraverse f = fmap Barbie . btraverse f . getBarbie
 
 
-instance (ProductBC b, AllBF Semigroup f b) => Semigroup (Barbie b f) where
+instance (ConstraintsB b, ApplicativeB b, AllBF Semigroup f b) => Semigroup (Barbie b f) where
   (<>) = bzipWith3 mk bdicts
     where
       mk :: Dict (ClassF Semigroup f) a -> f a -> f a -> f a
       mk = requiringDict (<>)
 
-instance (ProductBC b, AllBF Semigroup f b, AllBF Monoid f b) => Monoid (Barbie b f) where
+instance (ConstraintsB b, ApplicativeB b, AllBF Semigroup f b, AllBF Monoid f b) => Monoid (Barbie b f) where
   mempty  = bmempty
   mappend = (<>)